[MCP] shopify: address review feedback
